Move tests to impl
