+++ artifactory/commands/jobsummariesimpl/jobsummariesimpl_test.go
@@ -1,32 +1,31 @@
-package jobsummaries
+package jobsummariesimpl
 
 import (
 	buildinfo "github.com/jfrog/build-info-go/entities"
-	"github.com/jfrog/jfrog-cli-core/v2/artifactory/commands/jobsummariesimpl"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestBuildUiUrl(t *testing.T) {
-	gh := &jobsummariesimpl.GithubSummaryRtUploadImpl{
+	gh := &GithubSummaryRtUploadImpl{
 		PlatformUrl:     "https://myplatform.com/",
 		JfrogProjectKey: "myProject",
 	}
 	expected := "https://myplatform.com/ui/repos/tree/General/myPath/?projectKey=myProject"
-	actual := gh.BuildUiUrl("myPath")
+	actual := gh.buildUiUrl("myPath")
 	assert.Equal(t, expected, actual)
 
-	gh = &jobsummariesimpl.GithubSummaryRtUploadImpl{
+	gh = &GithubSummaryRtUploadImpl{
 		PlatformUrl:     "https://myplatform.com/",
 		JfrogProjectKey: "",
 	}
 	expected = "https://myplatform.com/ui/repos/tree/General/myPath/?projectKey="
-	actual = gh.BuildUiUrl("myPath")
+	actual = gh.buildUiUrl("myPath")
 	assert.Equal(t, expected, actual)
 }
 
 func TestBuildInfoTable(t *testing.T) {
-	gh := &jobsummariesimpl.GithubSummaryBpImpl{}
+	gh := &GithubSummaryBpImpl{}
 	gh.Builds = []*buildinfo.BuildInfo{
 		{
 			Name:     "buildName",
@@ -41,10 +40,10 @@ func TestBuildInfoTable(t *testing.T) {
 
 func TestParseBuildTime(t *testing.T) {
 	expected := "Jan 2, 2006 15:04:05"
-	actual := jobsummariesimpl.ParseBuildTime("2006-01-02T15:04:05.000-0700")
+	actual := parseBuildTime("2006-01-02T15:04:05.000-0700")
 	assert.Equal(t, expected, actual)
 
 	expected = "N/A"
-	actual = jobsummariesimpl.ParseBuildTime("")
+	actual = parseBuildTime("")
 	assert.Equal(t, expected, actual)
 }
